Log: Fix file path

+++ Src/RabbitCommonLog.h
@@ -82,8 +82,8 @@ private:
     QString m_szFileFormat;
     QFile m_File;
     QMutex m_Mutex;
-    quint64 m_nLength;   // Unit: byte
-    quint64 m_nCount;
+    qint64 m_nLength;   // Unit: byte
+    int m_nCount;
     QTimer m_Timer;
     #if QT_VERSION >= QT_VERSION_CHECK(5, 0, 0)
         static void myMessageOutput(QtMsgType type,
